[#67] refactor: 폴더 구조 변경(components)
