+++ src/app/(home)/components/GalaxyGraph.tsx
