+++ src/app/debateList/components/DebateListCard.tsx
