+++ src/app/newStar/components/NewStarForm.tsx
