+++ src/app/newStar/components/StarInput.tsx
