+++ src/app/newStar/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Wrapper from '@/components/Common/Wrapper';
-import NewStarForm from './NewStarForm/NewStarForm';
+import NewStarForm from './components/NewStarForm';
 
 // NOTE : 새 글쓰기 페이지
 const page = () => {
