+++ src/app/stars/[starId]/components/StarContent.tsx
