+++ src/app/stars/[starId]/components/StarContributor.tsx
