+++ src/app/stars/[starId]/components/StarInfo.tsx
